backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic
pydantic[email]==2.9.2
pydantic-settings==2.6.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.31.0
alembic==1.14.0

# Redis
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# HTTP Client
httpx==0.26.0

# Socket.IO
python-socketio==5.10.0

# Utilities
python-dotenv==1.0.0
